requirements.txt
fastapi
uvicorn
celery[redis]
sqlalchemy
asyncpg
alembic
pydantic
pydantic-settings
pymongo
motor
redis
httpx
feedparser
newspaper3k
spacy
transformers
torch
google-generativeai
tenacity
pandas
yfinance
praw
pytest
python-jose
python-dotenv
beautifulsoup4
